feat(weclawbot): v5.6 — drop SQLite, Postgres-only backend
最大幅简化。砍掉 SQLite 后端，daemon 永远跑 Postgres。直接消灭一类
反复制造 bug 的复杂度：

- v4→v5 升级时 refinery 兼容路径误标 V0009 为 applied → accounts 表缺
  platform_id → daemon 看 0 个账号 → 微信端"暂无法连接"
- v5.5 误部署 postgres feature binary → daemon 启动找不到 PG DSN → crash
- sqlx 0.8 不会自动 ? → $N 翻译（与之前 doc 描述相反），dual-backend
  必须 cfg-gate 76 个 callsite 或换 sqlx::Any，没有干净的中间路径

PG-only 后这些通通消失。

## 一次性数据迁移

新命令 `weclawbot import-sqlite ./state.db` 把存量 SQLite 数据迁进 PG。
本地用户的 ~/.weclawbot/state.db 已迁完: 2 accounts / 2 users /
49 user_history / 1 admin_key / 1 default tenant。SQLite 文件原封不动
留作 rollback。`rusqlite = "0.32"` 加入 deps 作为一次性 reader。

## PG-native migrations

V0001-V0009 用 PG 原生 DDL 重写：
- INTEGER PRIMARY KEY AUTOINCREMENT → BIGSERIAL PRIMARY KEY
- INTEGER 数值列 (message_count / count / version / msg_id) → BIGINT
  避免 i64 vs INT4 decode 错误
- REAL → DOUBLE PRECISION (避免 f64 vs FLOAT4)
- BLOB → BYTEA
- datetime('now') → to_char(now() AT TIME ZONE 'UTC', ...)

## ? → $1, $2 全代码改造

`scripts/pg_placeholder_rewrite.py` 一次性把 149 处 ? 占位符 (76 个
原本，加上后续新增的 query) 改成 PG 原生 $N，覆盖 15 个文件：
auth/admin_key.rs, cli/export_tenant.rs, ee/audit_archiver.rs,
repo/{accounts, admin_keys, audit, bindings, config_kv, dedup,
defaults, rate_limits, tenants, trust, users}_async.rs, service/admin.rs.

## 删除的代码

- `#[cfg(feature = "postgres")]` 所有分叉 (~80 行)
- `portable_ddl()` SQL 重写 (AUTOINCREMENT→BIGSERIAL 已 inline 到迁移文件)
- refinery 兼容 import (60 行) + V0009 self-heal (30 行)
- cli/backup.rs 双 backend 分叉 (50 行)
- PRAGMA foreign_keys / journal_mode / busy_timeout / synchronous 调用
- weclawbot-pg wrapper script (~/.local/bin)
- .github/workflows/postgres.yml (合并进 ci.yml)
- Cargo.toml 的 postgres feature flag

总共 ~400 行 SQLite-isms 删干净。

## PG-specific 修复

几处 SQLite 容忍 / PG 严格的差异手工修：
- `ON CONFLICT DO UPDATE SET count = count + 1` → `count = rate_limits.count + 1`
  (PG 不接受未限定的 column ref 在 SET 右侧)
- `SUM(count)` 返回 NUMERIC，需 `::BIGINT` cast 才能 decode 到 i64
- `SELECT 1` 返回 INT4，需 `SELECT 1::BIGINT`
- is_active CASE expression 返 INT4，需 `END::BIGINT`

## Test infrastructure

`open_in_memory()` 现在连 *_test 后缀的 DB 才允许（safety guard 防止
误删生产数据），每次调用前 `DROP SCHEMA public CASCADE; CREATE SCHEMA public`
重建 schema 模拟 SQLite :memory: 的 per-call 新鲜状态。CI 工作流
`POSTGRES_DB=weclawbot_test` 已对齐。

## CI

`postgres.yml` 删除（独立 PG matrix 不再需要）。ci.yml 加 postgres:16
service container 永远在跑，DATABASE_URL 注入到 cargo test。`--test-threads=1`
保留 (shared schema race 防御)。

## 验证

cargo test --release -- --test-threads=1 → **287 passed / 2 ignored**
cargo build --release --bin weclawbot → ✅
daemon 启动 + accounts:2 + healthz "db":"ok" + 微信端可通信 ✅

## 部署后操作

存量 SQLite 用户升级路径见 CLAUDE.md (新增 "Postgres backend" 段落):

  1. podman run -d --name weclawbot-pg ... postgres:16
  2. export WECLAWBOT_PG_URL=postgres://...
  3. weclawbot import-sqlite ~/.weclawbot/state.db
  4. weclawbot start

Idempotent 重跑安全。SQLite 文件原地保留作回滚。

+++ clawbnb/src/storage/migrations/V0003__dedup.sql
@@ -17,7 +17,7 @@
 -- window (~minutes) but small enough to keep table tiny.
 
 CREATE TABLE seen_messages (
-  msg_id         INTEGER PRIMARY KEY,
+  msg_id         BIGINT PRIMARY KEY,
   first_seen_at  TEXT NOT NULL
 );
